+++ lib/koala/http_service.rb
@@ -23,6 +23,23 @@ module Koala
       builder.adapter Faraday.default_adapter
     end
 
+    # Default servers for Facebook. These are read into the config OpenStruct,
+    # and can be overridden via Koala.config.
+    DEFAULT_SERVERS = {
+      :graph_server => 'graph.facebook.com',
+      :dialog_host => 'www.facebook.com',
+      :rest_server => 'api.facebook.com',
+      # certain Facebook services (beta, video) require you to access different
+      # servers. If you're using your own servers, for instance, for a proxy,
+      # you can change both the matcher and the replacement values.
+      # So for instance, if you're talking to fbproxy.mycompany.com, you could
+      # set up beta.fbproxy.mycompany.com for FB's beta tier, and set the
+      # matcher to /\.fbproxy/ and the beta_replace to '.beta.fbproxy'.
+      :host_path_matcher => /\.facebook/,
+      :video_replace => '-video.facebook',
+      :beta_replace => '.beta.facebook'
+    }
+
     # The address of the appropriate Facebook server.
     #
     # @param options various flags to indicate which server to use.